Add MacroExpander
MacroExpander is an IncAST version of iast's MacroProcessor.
This will make it easier to define macros to represent IncAST nodes
from Python source code, as well as macros that expand to multiple
IncAST nodes.

In the pre-Mars codebase, we had just used MacroProcessor directly
because we defined the IncAST language as an extension of Python's
nodes.