Remove expression contexts from IncAST
The ASDL is modified to delete the Read and Write nodes, as well as
the ctx field from Name, Tuple, and Attribute. IncAST logic and tests
are updated accordingly. Context is no longer considered when importing
Python nodes to IncAST. When exporting IncAST to Python, a filler value
of Load() is used uniformly for all contexts.

The reason for this change is that in most situations we don't care
whether we are reading or writing. In the places where we do care, we
can run some kind of visitor to determine this information for us.
Storing it on the node is unnecessary and makes code that deals with
these expressions more verbose.