Refine type lattice
Removed Union as a type because it's not clear to me how that fits
in with the lattice theory. It was only intended for the abstract
interpretation's well-typedness constraints, but that can be
implemented without Union by splitting into multiple separate
constraints.

Renamed subtype/supertype to smaller/bigger, since the lattice partial
order does not correspond exactly to domains of values.

Added eval_typestr().