Add QUERY directive
Added parsing of QUERY directive, but this information is not used at
the moment.

When parsing the directive, we make sure to convert the query source
code string to an expression AST early on, so that this AST can be
manipulated by subsequent rewritings. Otherwise, it's possible for a
rewriting to affect occurrences of the query without affecting the
corresponding QUERY directive, making things go out of sync. This
solution would fall apart anyway if a non-functional rewriting is
introduced, e.g. one that inserts fresh identifiers into queries.