Change SetFromMap semantics
The map value is now no longer a tuple of arbitrary arity, but rather
an arbitrary value that always corresponds to exactly one component of
the resulting relation. Updated runtime, transformation, tests.

This allows SetFromMap to be used on aggregate maintenance code without
having to make aggregate result maps map to singleton tuples.