Add uses_demand symbol attribute
This allows us to distinguish between no demand transformation, and
demand transformation where the U-set is nullary. Nullary U-sets are
important for demand-driven aggregates inside comprehensions, where
the aggregate takes no parameters. (Intuitively, it forces the
aggregate to have a value whenever the value is needed in the comp,
even if the operand is empty.)