Fix filtered clause renaming
Query variables in maintenance joins are renamed. This renaming should
also be applied to any filtered clauses that substitute clauses of the
maintenance join. The approach I use is a bit hackish in that it
relies on instrumenting the fresh name generators for joins, but this
averts the need to track such details in join.py.