Rearrange Unwrap logic
This eliminates use cases for QueryRewriter where the expand=True option
is used in a way that wraps the existing query node. This is important
for when we add annotations to Query nodes, since different occurrences
of query nodes may have different annotations but rewrite() is only
called on the first occurrence.

Added an Unwrapper helper, but had to use custom logic for comp
flattening since the unwrapping step is interleaved with rewriting
nested queries.