Change get_size_for_namespace() to explicitly only return top-level size
It gets the number of entries in a top-level set or auxiliary map, but
not the size of nested data that is accessible from these structures.
This is to avoid counting data that was present in the original program
(like celeb followers) when we intend to just count auxiliary structures
like dM_ub.

It may still count data in the original program under some cases, like
relational programs, but we can address that on a case-by-case basis.
The best solution in general is to measure the size for the
untransformed program and subtract that from the measurements for the
transformed programs.