Optimize ImgLookup implementation
Use "d[k] if k in d else v" paradigm instead of "d.get(k, v)", since
timeit tells us it's faster (by 2x if k is not in d).

Also use () instead of Set() for v when k is not in d, and it appears
on the RHS of a loop.